Add VaultHolder class for detection from other plugins
Prior to this commit plugins could not (easily) detect if an inventory a
player is using is a Player Vault inventory. This commit changes that by
assigning a "VaultHolder" to an inventory when it is created.

This new holder allows for plugins to easily detect if the inventory is
a player vault due to the new VaultHolder class.

The VaultHolder also holds information such as the vault number. The
vault number is required for further handling of the vault.

Without this addition plugins, such as AntiShare, cannot correctly
handle a vault.