Build for specific Bukkit release:
Eliminates awkward situations of changes causing confusion when suddenly
your plugin won't compile on new versions of Bukkit. Stuck to 1.7.9-R0.2
because this is the last release supporting getOnlinePlayers as an array,
which this plugin builds against, and building this way supports the most
servers, which is always an admirable goal. Wouldn't want to force people
onto the latest dev, would we?