Substantially reduce disk IO in mass destruction:
Previously, the signs config file was saved for every single change to its
contents. This works great until you have an episode of mass destruction.
Suddenly, you're getting the file saved multiple times per tick. Goodbye,
TPS. Hello, lag.

This commit replaces the functionality of the saveSigns method with
flagging the signs config as requiring a save. A scheduled task checking
for this boolean once per second allows for up-to-the-second accuracy of
the config file without substantially compromising server stability.

Lastly, the config is saved in onDisable to ensure the task hasn't missed
anything.