Move UUID caching code to convert command execution. Addresses #88
Instead of loading the cache up front (potentially taking minutes) we now
move the UUID caching to the command execution. Although this may increase
the command execution time it does reduce the time spent starting the
server and furthermore saves a little bit of memory from being used.

The cache is only seeded if the UUID provider has not been initialized. If
it has, then the seeding does not occur.